[섹션2] 훅 직접 만들기
- 예제 추가

#15-20201010
