+++ example/cra/src/Profile.jsx
@@ -1,11 +1,8 @@
-import React, {useState, useEffect} from 'react';
+import React from 'react';
+import useUser from './useUser';
 
 const Profile = ({userId}) => {
-  const [user, setUser] = useState(null);
-  useEffect(() => {
-    getUserAPI(userId).then(data => setUser(data));
-  }, [userId]);
-
+  const user = useUser(userId);
   return (
     <div>
       {!user && <p>사용자 정보를 가졍는 중...</p>}
@@ -17,17 +14,6 @@ const Profile = ({userId}) => {
       )}
     </div>
   )
-}
-
-const USER1 = {name: 'MJ', age: 20};
-const USER2 = {name: 'JH', age: 10};
-
-const getUserAPI = (userId) => {
-  return new Promise(res => {
-    setTimeout(() => {
-      res(userId % 2 ? USER1 : USER2);
-    }, 500)
-  })
 }
 
 export default Profile;
